test/mochitest/test_block_close.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1235459
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="http://mochi.test:8888/chrome/browser/extensions/b2g-installer/test/mochitest/common.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  </script>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<div id="container">
  <h3 id="offline"></h3>
  <ul id="devices"></ul>
  <input id="userBuild" type="file" value="Upload your own build" />
  <button id="installBtn">Flash Device</button>
  <div class="dialog" id="confirmDialog"><div>
    <section>
      <p>Your device has succesfully installed Firefox OS!</p>
       <button>Ok</button>
    </section>
  </div></div>
</div>
<pre id="test">
  <script type="application/javascript" id="code"></script>
  <script type="application/javascript">

  "use strict";

  function test_unload(aExpected) {
    return new Promise((resolve, reject) => {
      let beforeunloadEvent = new Event("beforeunload", 
                                        {"bubbles": false, "cancelable": true});
      window.dispatchEvent(beforeunloadEvent);
      console.debug("sending beforeunload event with", aExpected);
      setTimeout(() => {
        is(beforeunloadEvent.defaultPrevented, aExpected, "has been .preventDefault()");
	resolve();
      });
    });
  }

  function runTest() {
    loadAboutJs()
      .then(() => _isRisky = true)
      .then(() => test_unload(true))  // test that event is cancelled when risky
      .then(() => _isRisky = false)
      .then(() => test_unload(false)) // test it's not cancelled when not risky
      .then(() => SimpleTest.finish());
  }

  setPrefsAndRunTests();

  </script>
</pre>
</body>
</html>
